% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_histogram.R
\name{aggregate_histogram}
\alias{aggregate_histogram}
\title{Aggregate a histogram to remove categorical variables and sum to a single
waiting list}
\usage{
aggregate_histogram(histogram)
}
\arguments{
\item{histogram}{a waiting list in histogram format, possibly including
counts for different categorical variables}
}
\value{
a dataframe summarised to a single aggregated waiting list
}
\description{
Aggregate a histogram to remove categorical variables and sum to a single
waiting list
}
