% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_wait.R
\name{average_wait}
\alias{average_wait}
\title{Average Waiting Time}
\usage{
average_wait(target_wait, factor = 4)
}
\arguments{
\item{target_wait}{Numeric value of the number of weeks that has been set as
the target within which the patient should be seen.}

\item{factor}{Numeric factor used in average wait calculation - to get a
quarter of the target use factor=4 and one sixth of the target use factor =
6 etc. Defaults to 4.}
}
\value{
Numeric value of target mean waiting time to achieve a given target
wait.
}
\description{
This calculates the target mean wait given the two inputs of
target_wait and a numerical value for factor. The average wait is actually
the target mean wait and is calculated as follows: target_wait / factor. If
we want to have a chance between 1.8\%-0.2\% of making a waiting time target,
then the average patient should have a waiting time between a quarter and a
sixth of the target. Therefore: The mean wait should sit somewhere between
target_wait/factor=6 < Average Waiting Time < target_wait/factor=4.
}
\examples{
# If the target wait is 52 weeks then the target mean wait with a factor of 4
# would be 13 weeks and with a factor of 6 it would be 8.67 weeks.
average_wait(52, 4)
}
