% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_index.R
\name{calc_index}
\alias{calc_index}
\title{Calculate Column Indices}
\usage{
calc_index(waiting_list, colname = NULL, type = NULL)
}
\arguments{
\item{waiting_list}{a dataframe containing the waitinglist}

\item{colname}{string giving the column name}

\item{type}{if colname, write referral, withdrawal, removal to guess the index}
}
\value{
index
}
\description{
Internal Helper function to get column indicies for referrals,
removals, and withdrawals
}
