% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_priority_to_target.R
\name{calc_priority_to_target}
\alias{calc_priority_to_target}
\title{Calculates target days from priority code}
\usage{
calc_priority_to_target(priority)
}
\arguments{
\item{priority}{number 1,2,3 or 4}
}
\value{
number of days
}
\description{
Internal Helper function number of days from prioirty code
}
