% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tail_prob.R
\name{calc_target_capacity_from_target_probability}
\alias{calc_target_capacity_from_target_probability}
\title{target_capacity_from_target_probability}
\usage{
calc_target_capacity_from_target_probability(demand, target_prob, target_wait)
}
\arguments{
\item{demand}{dataframe. A df of referral dates and removals}

\item{target_prob}{numeric. The required waiting time}

\item{target_wait}{date. The start date to calculate from}
}
\value{
numeric
}
\description{
calculates a target capacity from a target probability#'
}
