% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bulk_synthetic_data.R
\name{create_bulk_synthetic_data}
\alias{create_bulk_synthetic_data}
\title{Create Bulk Synthetic Data}
\usage{
create_bulk_synthetic_data(bulk_data)
}
\arguments{
\item{bulk_data}{A dataframe object, each row being a waiting list with
parameters to generate the synthetic data. A sample data.frame is available
as demo-data}
}
\value{
Dataframe of waiting lists for each specified site and specialty,
opcs etc
}
\description{
Creates a series of waiting lists, one for each row in the dataframe
parameter and joins them together into one dataframe with relevant creation
criteria
}
\examples{
create_bulk_synthetic_data(demo_df)
}
