% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demographic.R
\docType{data}
\name{demographic_data}
\alias{demographic_data}
\title{demographic data}
\format{
An object of class \code{data.frame} with 5 rows and 9 columns.
}
\usage{
data(demographic_data)
}
\description{
demographic data
}
\examples{
\dontrun{data(demographic_data)}
}
\keyword{datasets}
