% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_snapshot_report.R
\name{make_snapshot_report}
\alias{make_snapshot_report}
\title{Make Waiting List Snapshot Report}
\usage{
make_snapshot_report(
  histogram_data,
  org_categories = NULL,
  fill_category = NULL,
  org_name = NULL,
  destination_directory = "."
)
}
\arguments{
\item{histogram_data}{data.frame.  A data.frame of histogram data in the NHSRwaitinglist data format.}

\item{org_categories}{charactor vector. A charactor vector of category column names present in the histogram data.}

\item{fill_category}{character. The name of a category column in the histogram data which will be used to define the histogram fill colours.}

\item{org_name}{character. The name of the organisation producing the report.}

\item{destination}{character. The path the final report should be rendered to.  Defaults to the current working directory.}
}
\value{
character. The filepath of the final rendered report.
}
\description{
Make Waiting List Snapshot Report
}
