% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_OPCS.R
\docType{data}
\name{OPCS4}
\alias{OPCS4}
\title{OPCS4 data}
\format{
Data frame with 9 columns
\describe{
\item{code_1digit}{The firt digit of the OPCS4 code, or 'chapter'}
\item{name_1digit}{The name/group of 'chapter' of the OPCS4 code}
\item{...}{Others to do with file is updated}
}
}
\source{
https://biobank.ndph.ox.ac.uk/ukb/coding.cgi?id=240
}
\usage{
data(OPCS4)
}
\description{
OPCS4 data
}
\examples{
\dontrun{
data(OPCS4)
}
}
\keyword{datasets}
