% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtt_improvement.R
\name{plot_rtt_journey}
\alias{plot_rtt_journey}
\title{Create a ggplot of an RTT improvement journey}
\usage{
plot_rtt_journey(referral_rate)
}
\arguments{
\item{referral_rate}{numeric. A weekly referral rate to plot}
}
\value{
ggplot object
}
\description{
Create a ggplot of an RTT improvement journey
}
