% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue_scheduler_DELETE.R
\name{queue_scheduler}
\alias{queue_scheduler}
\title{A simple operation scheduler}
\usage{
queue_scheduler(waiting_list, schedule)
}
\arguments{
\item{waiting_list}{dataframe consisting addition and removal dates}

\item{schedule}{a list of dates}
}
\value{
updated waiting list with feasible dates schedul
}
\description{
Takes a list of dates and schedules them to a waiting list.
This is done in date order. I.e. earlier referrals are scheduled first.
}
\examples{
referrals <- c.Date("2024-01-01","2024-01-04","2024-01-10","2024-01-16")
removals <- c.Date("2024-01-08",NA,NA,NA)
waiting_list <- data.frame("referral" = referrals ,"removal" = removals )
schedule <- c.Date("2024-01-03","2024-01-05","2024-01-18")
updated_list <- queue_scheduler(waiting_list,schedule)
}
