% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtt_improvement.R
\name{rep_rtt_improvement}
\alias{rep_rtt_improvement}
\title{Make an RTT improvement report}
\usage{
rep_rtt_improvement(referral_rates)
}
\arguments{
\item{referral_rates}{numeric. A numeric vector of referral rates. Each rate
will produce a plot in the finished report}
}
\value{
character. The path to a produced report
}
\description{
Make an RTT improvement report
}
