% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exponential_histogram.R
\name{sim_exponential_histogram}
\alias{sim_exponential_histogram}
\title{Create a simulated histogram in the format used by the package}
\usage{
sim_exponential_histogram(
  num_intervals = 52,
  end_date = Sys.Date(),
  rate = 0.1,
  queue_size = 1000,
  time_interval = "weeks",
  random = FALSE
)
}
\arguments{
\item{num_intervals}{integer. Number of time intervals to create.}

\item{end_date}{Date. The date that the waiting list should end.}

\item{rate}{numeric. The rate defining the exponential.}

\item{queue_size}{numeric. Total queue size used to scale the histogram.}

\item{time_interval}{Character or numeric. Controls the spacing between rows.}

\item{random}{logical. If TRUE, draws interval counts from an exponential process.}
}
\value{
A data.frame of a simulated waiting list in histogram format.
}
\description{
Create a simulated histogram in the format used by the package
}
