% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_schedule.R
\name{sim_schedule}
\alias{sim_schedule}
\title{Generator a list of dates to schedule}
\usage{
sim_schedule(n_rows = 10, start_date = NULL, daily_capacity = 1)
}
\arguments{
\item{n_rows}{Number of rows/patients to generate}

\item{start_date}{Start date (needed to generate patient ages)}

\item{daily_capacity}{Number of patients per day}
}
\value{
data.frame. Empty waiting list.
}
\description{
Generates a list if dates in a given range
}
