% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tail_prob.R
\name{target_capacity_from_target_probability}
\alias{target_capacity_from_target_probability}
\title{target_capacity_from_target_probability}
\usage{
target_capacity_from_target_probability(demand, target_prob, target_wait)
}
\arguments{
\item{demand}{dataframe. A df of referral dates and removals}

\item{target_prob}{numeric. The required waiting time}

\item{target_wait}{date. The start date to calculate from}
}
\description{
target_capacity_from_target_probability
}
