% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_plot_hist.R
\name{vis_plot_hist}
\alias{vis_plot_hist}
\title{Plot Actual vs Geometric Target Waiting List Histogram}
\usage{
vis_plot_hist(wl_hist, target = 18, percentage = 92)
}
\arguments{
\item{wl_hist}{Data frame containing waiting list history.}

\item{target}{Numeric. Target number of weeks for the geometric distribution (e.g., 18).}

\item{percentage}{Numeric. Percentile to highlight (e.g., 92).}
}
\value{
A ggplot object showing the actual and target distributions.
}
\description{
This function plots the actual waiting list histogram against a geometric target distribution.
}
