% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_insert_hist.R
\name{wl_insert_hist}
\alias{wl_insert_hist}
\title{Insert some patients into a waiting list in histogram format}
\usage{
wl_insert_hist(waiting_list, additions)
}
\arguments{
\item{waiting_list}{dataframe. A waiting list in histogram format}

\item{additions}{dataframe. A dataframe of waiting list additions in
histogram format}
}
\value{
data.frame. A combined waiting list.
}
\description{
Acting on a histogram this is identical to \\code{wl_join_hist()}.
}
