% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_mean_wait_age.R
\name{wl_mean_wait_age}
\alias{wl_mean_wait_age}
\title{Calculate Mean Age of Waiting for Patients on Waiting List}
\usage{
wl_mean_wait_age(waiting_list, referral_index, removal_index, end_date)
}
\arguments{
\item{waiting_list}{A data frame containing patient waiting list data.}

\item{removal_index}{The column name or index for the removal date in \code{waiting_list}.}

\item{end_date}{The date (as Date or numeric) up to which to calculate ages.}

\item{dob_index}{The column name or index for the date of birth in \code{waiting_list}.}
}
\value{
The mean age (numeric) for patients still waiting as of \code{end_date}.
}
\description{
This function calculates the time waiting for patients who are still on the waiting list as of a specified end date.
Recall that the age of waiting is the mean time waited so far. (Opposed to the mean sojourn time, which is the time waiting until removal from the list.)
}
