% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_queue_size_hist.R
\name{wl_queue_size_hist}
\alias{wl_queue_size_hist}
\title{Calculates the number of people in a queue}
\usage{
wl_queue_size_hist(wl_hist)
}
\arguments{
\item{wl_hist}{dataframe. A waiting list in histogram format}
}
\value{
integer. The size of the queue or waiting list
}
\description{
Calculates the number of people in a queue
}
