% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_referral_stats.R
\name{wl_referral_stats}
\alias{wl_referral_stats}
\title{Calculate some stats about referrals}
\usage{
wl_referral_stats(waiting_list, start_date = NULL, end_date = NULL)
}
\arguments{
\item{waiting_list}{dataframe. A df of referral dates and removals}

\item{start_date}{date. The start date to calculate from}

\item{end_date}{date. The end date to calculate to}
}
\value{
dataframe. A df containing number of referrals, mean demand, and the coefficient of variation of referrals
}
\description{
Calculate some stats about referrals
}
\examples{
referrals <- c.Date("2024-01-01", "2024-01-04", "2024-01-10", "2024-01-16")
removals <- c.Date("2024-01-08", NA, NA, NA)
waiting_list <- data.frame("referral" = referrals , "removal" = removals)
referral_stats <- wl_referral_stats(waiting_list)
}
