% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_referral_stats_hist.R
\name{wl_referral_stats_hist}
\alias{wl_referral_stats_hist}
\title{Calculate some stats about referrals from histogram data}
\usage{
wl_referral_stats_hist(
  wl_hist,
  start_date = NULL,
  end_date = NULL,
  time_interval = "weeks"
)
}
\arguments{
\item{wl_hist}{data.frame. A histogram with columns: arrival_since, arrival_before,
n, and report_date.}

\item{start_date}{Date or character (in format 'YYYY-MM-DD'); The start date
to calculate from.}

\item{end_date}{Date or character (in format 'YYYY-MM-DD'); The end date to
calculate to.}

\item{time_interval}{Character or numeric. Passed through to format_histogram() when
deriving the histogram end date.}
}
\value{
A data.frame with the following summary statistics on
referrals/demand:

\describe{
\item{demand_weekly}{Numeric. Mean number of additions to the waiting list
per week.}
\item{demand_daily}{Numeric. Mean number of additions to the waiting list
per day.}
\item{demand_cov}{Numeric. Coefficient of variation in the time between
additions to the waiting list.}
\item{demand_count}{Numeric. Total demand over the full time period.}
}
}
\description{
Calculate referral statistics from a histogram containing multiple report dates.
}
\examples{
referrals <- as.Date(c("2024-01-01", "2024-01-04", "2024-01-10", "2024-01-16"))
removals <- as.Date(c("2024-01-08", NA, NA, NA))
hist_waiting_list <- data.frame(
  arrival_since = as.Date(c("2024-01-01", "2024-01-08")),
  arrival_before = as.Date(c("2024-01-07", "2024-01-14")),
  n = c(100, 120),
  report_date = as.Date(c("2024-01-31", "2024-02-29"))
)
referral_stats <- wl_referral_stats_hist(hist_waiting_list)

}
