% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_big_simulator.R,
%   R/wl_simulator-DA-H12JD099Q6X3.R, R/wl_simulator.R
\name{wl_simulator}
\alias{wl_simulator}
\title{Simple simulator to create a waiting list}
\usage{
wl_simulator(
  start_date = NULL,
  end_date = NULL,
  demand = 10,
  capacity = 11,
  waiting_list = NULL,
  referral_index = 1,
  withdrawal_prob = NA,
  detailed_sim = FALSE
)

wl_simulator(
  start_date = NULL,
  end_date = NULL,
  demand = 10,
  capacity = 11,
  waiting_list = NULL,
  referral_index = 1,
  withdrawal_prob = NA,
  detailed_sim = FALSE
)

wl_simulator(
  start_date = NULL,
  end_date = NULL,
  demand = 10,
  capacity = 11,
  waiting_list = NULL,
  referral_index = 1,
  withdrawal_prob = NA,
  detailed_sim = FALSE
)
}
\arguments{
\item{start_date}{date. The start date for the simulation}

\item{end_date}{date. The end date for the simulation}

\item{demand}{numeric. Weekly demand (ie. typical referrals per week)}

\item{capacity}{numeric. Weekly capacity (ie. typical removals per week)}

\item{waiting_list}{integer. The number of patients on the waiting list}

\item{referral_index}{integer. The column number in the waiting_list which
contains the referral dates}
}
\value{
dataframe. A df of simulated referrals and removals

dataframe. A df of simulated referrals and removals

dataframe. A df of simulated referrals and removals
}
\description{
Creates a simulated waiting list comprising referral dates,
and removal dates

Creates a simulated waiting list comprising referral dates,
and removal dates

Creates a simulated waiting list comprising referral dates,
and removal dates
}
\examples{

over_capacity_simulation <-
  wl_simulator("2024-01-01", "2024-03-31", 100, 110)
under_capacity_simulation <-
  wl_simulator("2024-01-01", "2024-03-31", 100, 90)

# TODO
# error messages (e.g. start_date > end_date)

over_capacity_simulation <-
  wl_simulator("2024-01-01", "2024-03-31", 100, 110)
under_capacity_simulation <-
  wl_simulator("2024-01-01", "2024-03-31", 100, 90)

# TODO
# error messages (e.g. start_date > end_date)

over_capacity_simulation <-
  wl_simulator("2024-01-01", "2024-03-31", 100, 110)
under_capacity_simulation <-
  wl_simulator("2024-01-01", "2024-03-31", 100, 90)

# TODO
# error messages (e.g. start_date > end_date)
}
