% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_stats.R
\name{wl_stats}
\alias{wl_stats}
\title{Waiting List Statistics}
\arguments{
\item{waiting_list}{dataframe consisting addition and removal dates}

\item{target_wait}{the required waiting time}
}
\value{
summary of key waiting list statistics
}
\description{
A summary of all the key statistics associated with a waiting list
}
\examples{

TO DO!!
MAKE CONSISTENT NOTATION
default start and end date if empty
}
