% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_stats_categories.R
\name{wl_stats_categories}
\alias{wl_stats_categories}
\title{Calculate some stats about the waiting list}
\usage{
wl_stats_categories(
  waiting_list,
  target_wait = 4,
  categories = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{waiting_list}{dataframe. A df of referral dates and removals}

\item{target_wait}{numeric. The required waiting time}

\item{start_date}{date. The start date to calculate from}

\item{end_date}{date. The end date to calculate to}
}
\value{
dataframe. A df of important waiting list statistics
}
\description{
A summary of all the key stats associated with a waiting list
}
\examples{

referrals <- c.Date("2024-01-01", "2024-01-04", "2024-01-10", "2024-01-16")
removals <- c.Date("2024-01-08", NA, NA, NA)
waiting_list <- data.frame("referral" = referrals, "removal" = removals)
waiting_list_stats <- wl_stats(waiting_list)

# TO DO!!
# Check target_capacity weekly daily
# Error if dates are in the wrong order
# Calculate the number of missed operations
# Start date and end date calculations not working well
#
# MAKE CONSISTENT NOTATION
# default start and end date if empty
# make units of output weekly operations not daily
# mean removal too big
# Z score in capacity calculations
# mean demand and mean capacity not interrarraival and departures plesae.
}
