% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_histogram.R
\name{aggregate_histogram}
\alias{aggregate_histogram}
\title{Aggregate a histogram to remove categorical variables and sum to a single
waiting list}
\usage{
aggregate_histogram(histogram, group_columns = NULL)
}
\arguments{
\item{histogram}{a waiting list in histogram format, possibly including
counts for different categorical variables}
}
\value{
a dataframe summarised to a single aggregated waiting list
}
\description{
(arrival_since, arrival_before) are the start and end dates of the arrival time
interval. The function aggregates the histogram by summing the counts.
(The arrival_since and arrival_before columns are not aggregated.)
}
