% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demographic.R
\docType{data}
\name{demographic_data}
\alias{demographic_data}
\title{demographic data}
\format{
Data frame with 9 columns
\describe{
\item{hospital_site}{ODS hospital site code}
\item{...}{Others to do with file is updated}
}
}
\usage{
data(demographic_data)
}
\description{
demographic data
}
\examples{

data(demographic_data)

}
\keyword{datasets}
