% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPCS_data.R
\docType{data}
\name{OPCS4}
\alias{OPCS4}
\title{Arabidopsis QTL data on gravitropism}
\format{
An object of class \code{"cross"}; see \code{\link[qtl:read.cross]{qtl::read.cross()}}.
}
\source{
Mouse Phenome Database, \url{https://phenome.jax.org/projects/Moore1b}
}
\usage{
data(OPCS4)
}
\description{
Data from a QTL experiment on gravitropism in
Arabidopsis, with data on 162 recombinant inbred lines (Ler x
Cvi). The outcome is the root tip angle (in degrees) at two-minute
increments over eight hours.
}
\examples{
data(OPCS4)
}
\references{
Moore CR, Johnson LS, Kwak I-Y, Livny M, Broman KW,
Spalding EP (2013) High-throughput computer vision introduces the
time axis to a quantitative trait map of a plant growth
response. Genetics 195:1077-1086
(\href{https://pubmed.ncbi.nlm.nih.gov/23979570/}{PubMed})
}
\keyword{datasets}
