% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPCS_data.R
\docType{data}
\name{OPCS4}
\alias{OPCS4}
\title{OPCS4 data}
\format{
An object of class \code{data.frame} with 12279 rows and 9 columns.
}
\source{
https://biobank.ndph.ox.ac.uk/ukb/coding.cgi?id=240
}
\usage{
data(OPCS4)
}
\description{
OPCS4 data
}
\examples{
data(OPCS4)
}
\keyword{datasets}
