% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtt_improvement.R
\name{plot_rtt_journey}
\alias{plot_rtt_journey}
\title{Create a ggplot of an RTT improvement journey}
\usage{
plot_rtt_journey(
  referral_rate,
  rtt_performance = c(0.5, 0.6, 0.7, 0.8, 0.92),
  n_bins = 64
)
}
\arguments{
\item{referral_rate}{numeric. A weekly referral rate to plot}

\item{rtt_performance}{numeric vector. A vector of performance thresholds, as proportion e.g. 0.65 = 65\%.}

\item{n_bins}{numeric.  Number of bins for histogram. Default is 1 - 64 weeks.}
}
\value{
ggplot object
}
\description{
Create a ggplot of an RTT improvement journey
}
