% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queue_scheduler.R
\name{queue_scheduler}
\alias{queue_scheduler}
\title{schedules removal dates on a waiting list}
\usage{
queue_scheduler(waiting_list, schedule)
}
\arguments{
\item{waiting_list}{dataframe}

\item{schedule}{Numeric value}
}
\value{
updated_list
}
\description{
schedules removal dates on a waiting list
}
