% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exponential_histogram.R
\name{sim_exponential_histogram}
\alias{sim_exponential_histogram}
\title{Create a weekly histogram in the format used by the package}
\usage{
sim_exponential_histogram(weeks = 52, end_date = Sys.Date(), rate = 0.1)
}
\arguments{
\item{weeks}{integer. Number of weeks of waiting list to create}

\item{end_date}{data. The date that the waiting list should end}

\item{rate}{numeric. The rate defining the exponential}
}
\value{
a dataframe of a simulated waiting list in histogram format
}
\description{
Create a weekly histogram in the format used by the package
}
