% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_exponential_histogram.R
\name{sim_exponential_histogram}
\alias{sim_exponential_histogram}
\title{Create a waiting list histogram in the format used by the package}
\usage{
sim_exponential_histogram(
  num_intervals = 52,
  end_date = Sys.Date(),
  rate = 0.1,
  queue_size = 1000,
  time_interval = "weeks",
  random = FALSE
)
}
\arguments{
\item{num_intervals}{integer. Number of intervals (weeks, months, etc) of waiting list to create}

\item{end_date}{data. The date that the waiting list should end}

\item{rate}{numeric. The rate defining the exponential}

\item{queue_size}{integer. The queue size to create}

\item{time_interval}{character. The bin width for the histogram (weeks, or months)}

\item{random}{logical. Whether to generate a random histogram (TRUE/FALSE)}
}
\value{
a data.frame of a simulated waiting list in histogram format
}
\description{
Create a waiting list histogram in the format used by the package
}
