% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_patients.R
\name{sim_patients}
\alias{sim_patients}
\title{Generator of NHS patients}
\usage{
sim_patients(n_rows = 10, start_date = NULL)
}
\arguments{
\item{n_rows}{Number of rows/patients to generate}

\item{start_date}{Start date (needed to generate patient ages)}
}
\value{
dataframe. Empty waiting list.
}
\description{
Generator of NHS patients
}
\examples{
sim_patients(100,"2024-09-21")
}
