% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_join_hist.R
\name{wl_join_hist}
\alias{wl_join_hist}
\title{Join two waiting lists provided as histograms}
\usage{
wl_join_hist(wl_hist_1, wl_hist_2)
}
\arguments{
\item{wl_hist_1}{dataframe. A waiting list in histogram format}

\item{wl_hist_2}{dataframe. A waiting list in histogram format}
}
\value{
dataframe. A combined waiting list in histogram format
}
\description{
Join two waiting lists provided as histograms
}
