% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_percentile_hist.R
\name{wl_percentile_hist}
\alias{wl_percentile_hist}
\title{Calculate the Date and Weeks Corresponding to a Given Percentile in Waiting List History}
\usage{
wl_percentile_hist(wl_hist, percentage = 92)
}
\arguments{
\item{wl_hist}{A data frame containing waiting list history with columns: n (number of patients), arrival_since, arrival_before, and report_date.}

\item{percentage}{Numeric value indicating the percentile to calculate (default is 92).}
}
\value{
A list with:
\item{date_percentile}{The estimated date corresponding to the given percentile.}
\item{weeks_percentile}{The number of weeks from the report date to the percentile date.}
}
\description{
Calculate the Date and Weeks Corresponding to a Given Percentile in Waiting List History
}
\examples{
# get_percentile_date(wl_hist, percentage = 92)
}
