% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_referral_stats.R
\name{wl_referral_stats}
\alias{wl_referral_stats}
\title{insert waiting list additions}
\usage{
wl_referral_stats(waiting_list, start_date = NULL, end_date = NULL)
}
\arguments{
\item{waiting_list}{a data frame of referral dates and removals}

\item{additions}{a list of referral dates}
}
\value{
an updated waiting list
}
\description{
adds new referrals (removal date is set NA)
}
\examples{
> referrals <- c.Date("2024-01-01","2024-01-04","2024-01-10","2024-01-16")
> removals <- c.Date("2024-01-08",NA,NA,NA)
> waiting_list <- data.frame("referral" = referrals ,"removal" = removals )
> additions <- c.Date("2024-01-03","2024-01-05","2024-01-18")
> wl_insert(waiting_list, additions)
}
