% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_referral_stats_hist-E-LOSX4GYQ6L4.R,
%   R/wl_referral_stats_hist.R
\name{wl_referral_stats_hist}
\alias{wl_referral_stats_hist}
\title{Calculate some stats about referrals}
\usage{
wl_referral_stats_hist(
  wl_hist,
  start_date = NULL,
  end_date = NULL,
  time_interval = "weeks"
)

wl_referral_stats_hist(
  wl_hist,
  start_date = NULL,
  end_date = NULL,
  time_interval = "weeks"
)
}
\arguments{
\item{start_date}{Date or character (in format 'YYYY-MM-DD'); The start date
to calculate from}

\item{end_date}{Date or character (in format 'YYYY-MM-DD'); The end date to
calculate to}

\item{waiting_list}{data.frame. A df of referral dates and removals}

\item{referral_index}{the column index of referrals}
}
\value{
A data.frame with the following summary statistics on
referrals/demand:

\describe{
\item{demand_weekly}{Numeric. Mean number of additions to the waiting list
per week.}
\item{demand_daily}{Numeric. Mean number of additions to the waiting list
per day.}
\item{demand_cov}{Numeric. Coefficient of variation in the time between
additions to the waiting list.}
\item{demand_count}{Numeric. Total demand over the full time period.}
}

A data.frame with the following summary statistics on
referrals/demand:

\describe{
\item{demand_weekly}{Numeric. Mean number of additions to the waiting list
per week.}
\item{demand_daily}{Numeric. Mean number of additions to the waiting list
per day.}
\item{demand_cov}{Numeric. Coefficient of variation in the time between
additions to the waiting list.}
\item{demand_count}{Numeric. Total demand over the full time period.}
}
}
\description{
Calculate some stats about referrals

Calculate some stats about referrals
}
\examples{
referrals <- c.Date("2024-01-01", "2024-01-04", "2024-01-10", "2024-01-16")
removals <- c.Date("2024-01-08", NA, NA, NA)
waiting_list <- data.frame("referral" = referrals, "removal" = removals)
referral_stats <- wl_referral_stats(waiting_list)

referrals <- c.Date("2024-01-01", "2024-01-04", "2024-01-10", "2024-01-16")
removals <- c.Date("2024-01-08", NA, NA, NA)
waiting_list <- data.frame("referral" = referrals, "removal" = removals)
referral_stats <- wl_referral_stats(waiting_list)

}
