% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_schedule.R
\name{wl_schedule}
\alias{wl_schedule}
\title{schedules removal dates on a waiting list}
\usage{
wl_schedule(waiting_list, schedule, referral_index = 1, removal_index = 2)
}
\arguments{
\item{waiting_list}{dataframe}

\item{schedule}{Numeric value}
}
\value{
updated_list
}
\description{
schedules removal dates on a waiting list
}
