% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_simulator.R
\name{wl_simulator}
\alias{wl_simulator}
\title{Simple Simulator}
\usage{
wl_simulator(
  start_date,
  end_date,
  demand,
  capacity,
  waiting_list = NULL,
  referral_index = 1
)
}
\arguments{
\item{start_date}{Start date}

\item{end_date}{End date}

\item{demand}{Demand}

\item{capacity}{Capacity}
}
\value{
updated_list a new waiting list
}
\description{
Simulators
}
\examples{

#TODO error messages (e.g. start_date > end_date)
}
