% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_simulator.R
\name{wl_simulator}
\alias{wl_simulator}
\title{Simple Simulator}
\usage{
wl_simulator(start_date, end_date, demand, capacity)
}
\arguments{
\item{capacity}{}
}
\value{
updated_list a new waiting list
}
\description{
Simulators
}
