% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wl_stats.R
\name{wl_stats}
\alias{wl_stats}
\title{Waiting List Statistics}
\usage{
wl_stats(waiting_list, target_wait = 4, start_date = NULL, end_date = NULL)
}
\arguments{
\item{waiting_list}{dataframe consisting addition and removal dates}

\item{target_wait}{the required waiting time}
}
\value{
summary of key waiting list statistics
}
\description{
A summary of all the key statistics associated with a waiting list
}
\examples{

# TO DO!!
# Check target_capacity weekly daily
# Error if dates are in the wrong order
# Calculate the number of missed operations
# Start date and end date calculations not working well
# MAKE CONSISTENT NOTATION
# default start and end date if empty
# make units of output weekly operations not daily
# mean removal too big
# Z score in capacity calculations
# mean demand and mean capacity not interrarraival and departures plesae.
}
